// backend/src/ts/model/result-raw.ts
export class ResultRaw {
    saison: number;
    discipline: string;
    licence: string;
    nom: string;
    prenom: string;
    horsF: string;
    sexe: string;
    categorie: string;
    categorieS: string;
    codeStructure: string;
    nomStructure: string;
    arme: string;
    niveau: string;
    score: number;
    paille: number;
    dix: number;
    neuf: number;
    distance: number;
    blason: string;
    dateDebutConcours: string;
    dateFinConcours: string;
    lieuConcours: string;
    codeStructureOrganisatrice: string;
    nomStructureOrganisatrice: string;
    formuleTir: string;
    niveauChpt: string;
    detailNiveauChpt: string;
    distinction: string;
    placeQualif: number;
    scoreDist1: number;
    scoreDist2: number;
    scoreDist3: number;
    scoreDist4: number;
    score32: number;
    score16: number;
    score8: number;
    scoreQuart: number;
    scoreDemi: number;
    scorePetiteFinal: number;
    scoreFinal: number;
    placeDef: number;
    numDepart: number;
    eprvNom: string;
    catTir: string;
    catClass: string;
    scoreValide: string;

    constructor(row: string[]) {
        this.saison = parseInt(row[0]);
        this.discipline = row[1];
        this.licence = row[2];
        this.nom = row[3];
        this.prenom = row[4];
        this.horsF = row[5];
        this.sexe = row[6];
        this.categorie = row[7];
        this.categorieS = row[8];
        this.codeStructure = row[9];
        this.nomStructure = row[10];
        this.arme = row[11];
        this.niveau = row[12];
        this.score = parseInt(row[13]);
        this.paille = parseInt(row[14]);
        this.dix = parseInt(row[15]);
        this.neuf = parseInt(row[16]);
        this.distance = parseInt(row[17]);
        this.blason = row[18];
        this.dateDebutConcours = row[19];
        this.dateFinConcours = row[20];
        this.lieuConcours = row[21];
        this.codeStructureOrganisatrice = row[22];
        this.nomStructureOrganisatrice = row[23];
        this.formuleTir = row[24];
        this.niveauChpt = row[25];
        this.detailNiveauChpt = row[26];
        this.distinction = row[27];
        this.placeQualif = parseInt(row[28]);
        this.scoreDist1 = parseInt(row[29]);
        this.scoreDist2 = parseInt(row[30]);
        this.scoreDist3 = parseInt(row[31]);
        this.scoreDist4 = parseInt(row[32]);
        this.score32 = parseInt(row[33]);
        this.score16 = parseInt(row[34]);
        this.score8 = parseInt(row[35]);
        this.scoreQuart = parseInt(row[36]);
        this.scoreDemi = parseInt(row[37]);
        this.scorePetiteFinal = parseInt(row[38]);
        this.scoreFinal = parseInt(row[39]);
        this.placeDef = parseInt(row[40]);
        this.numDepart = parseInt(row[41]);
        this.eprvNom = row[42];
        this.catTir = row[43];
        this.catClass = row[44];
        this.scoreValide = row[45];
    }

}